+++ src/lib/misc/errors.ts
@@ -1,0 +1,5 @@
+export class UguisuError extends Error {
+	constructor(message: string) {
+		super(message);
+	}
+}
