+++ src/lib/misc/options.ts
@@ -1,0 +1,4 @@
+export type UguisuOptions = {
+	stdin?: () => Promise<string>,
+	stdout?: (buf: string) => void,
+};
