+++ src/lib/scan.ts
@@ -1,4 +1,4 @@
-import { UguisuError } from './index.js';
+import { UguisuError } from './misc/errors.js';
 import { Trace } from './misc/trace.js';
 
 const trace = Trace.getDefault().createChild(false);
